fix(ci): python full coverage - isolate heartbeat readiness side effects
